Fix List Deallocation When Allocation Is Adjacent To Next Chunk
Thorougher testing had revealed the test case mentioned in the commit
title was non functional. The bug was related to the mismatch in size
between the allocated chunk struct and free chunk struct. The allocated
chunk struct requires more size and thus there needs to be special
handling when converting a free chunk struct to an allocated chunk
struct.